My first project in university.
C programming. Aiming to implement a naive
DBMS, support very limited SQL query.

I wrote another version to one of my classmates
and implemented the data representation with Linked list.

My version does it by array, which is easier to implment and
efficicy is rather good.

Don't know Flex and Bison, maybe Lex and Yacc. So wrote it
with manual if-else.